further organized subapp resources
