+++ src/app.module.ts
@@ -19,7 +19,7 @@ import { APP_GUARD } from '@nestjs/core';
 import { RolesGuard } from './iam/authorization/guards/roles.guard';
 import { AuthenticationGuard } from './iam/authentication/guards/authentication/authentication.guard';
 import { AccessTokenGuard } from './iam/authentication/guards/access-token/access-token.guard';
-import { SubappsModule } from './subapps/subapps.module';
+import { SubappsModule } from './subapps/resources/subapps.module';
 
 @Module({
   imports: [
