+++ src/subapps/resources/subapps.controller.ts
@@ -10,6 +10,8 @@ import { SubappsService } from './subapps.service';
 import { Auth } from 'src/iam/decorators/auth.decorator';
 import { AuthType } from 'src/iam/enums/auth-type.enum';
 import { FileInterceptor } from '@nestjs/platform-express';
+
+// REMOVE AuthType.None AND APPLY PROPER AUTHORIZATION FOR ALL SUBAPPS!!!
 @Auth(AuthType.None)
 @Controller('subapps')
 export class SubappsController {
