+++ src/subapps/resources/subapps.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { SubappsService } from './subapps.service';
 import { SubappsController } from './subapps.controller';
-import { FreeinvModule } from './freeinv/freeinv.module';
+import { FreeinvModule } from '../freeinv/freeinv.module';
 
 @Module({
   controllers: [SubappsController],
